# DormShareAPI/Documentation/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/item/details/{itemId}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Post Details",
        "operationId": "post_Details_item_details__itemId__get",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Itemid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemReadWithImages"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/item/category/{category}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Categorized Feed",
        "operationId": "categorized_Feed_item_category__category__get",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemReadWithImages"
                  },
                  "title": "Response Categorized Feed Item Category  Category  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/item/post": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Post Item",
        "operationId": "post_Item_item_post_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/item/status/update/{itemId}": {
      "patch": {
        "tags": [
          "Items"
        ],
        "summary": "Update Itemstatus",
        "operationId": "Update_ItemStatus_item_status_update__itemId__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Itemid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemReadWithImages"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/item/update/{itemId}": {
      "patch": {
        "tags": [
          "Items"
        ],
        "summary": "Update Item",
        "operationId": "update_Item_item_update__itemId__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Itemid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemReadWithImages"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/item/delete": {
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Post Delete",
        "operationId": "post_Delete_item_delete_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Itemid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "photo_url": {
            "type": "string",
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "photo_url"
        ],
        "title": "ImageRead"
      },
      "ItemCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 75,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          },
          "trade_type": {
            "type": "string",
            "maxLength": 70,
            "minLength": 1,
            "title": "Trade Type"
          },
          "price": {
            "type": "string",
            "minLength": 1,
            "title": "Price"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "title": "Category"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "trade_type",
          "price",
          "category"
        ],
        "title": "ItemCreate"
      },
      "ItemReadWithImages": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "trade_type": {
            "type": "string",
            "title": "Trade Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageRead"
            },
            "type": "array",
            "title": "Images",
            "default": []
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewRead"
            },
            "type": "array",
            "title": "Reviews",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "price",
          "trade_type",
          "category",
          "is_available",
          "created_at",
          "owner_id"
        ],
        "title": "ItemReadWithImages"
      },
      "ItemUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "title": "ItemUpdate"
      },
...
      "ReviewRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "borrower_id": {
            "type": "string",
            "format": "uuid",
            "title": "Borrower Id"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "stars_amount": {
            "type": "number",
            "title": "Stars Amount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "borrower_id",
          "transaction_id",
          "text",
          "stars_amount"
        ],
        "title": "ReviewRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "auth/login-swagger"
          }
        }
      }
    }
  }
}